requirements/requirements.txt
appdirs==1.4.3
attrs==19.3.0
cached-property==1.5.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
defusedxml==0.6.0
Flask==1.1.2
idna==2.9
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.5.0
MarkupSafe==1.1.1
mysql-connector==2.2.9
numpy==1.18.4
pandas==1.0.3
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
requests-toolbelt==0.9.1
six==1.14.0
urllib3==1.25.9
Werkzeug==1.0.1
zeep==3.4.0